refactor: mv default props to calendar picker
